64-bit warnings cleared
git-svn-id: svn://orca.physics.unc.edu/Orca64@11907 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/SNOP/SNOPController.m
@@ -2481,7 +2481,7 @@ snopGreenColor;
             break;
     }
     @try{
-        units = (int)[self decideUnitsToUseForRow:activeCell usingModel:mtcModel];
+        units = (int)[self decideUnitsToUseForRow:(int)activeCell usingModel:mtcModel];
         [mtcModel setThresholdOfType:threshold_index fromUnits:units toValue:threshold_value];
     }
     @catch(NSException *excep) {
